+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/FrameworkExtensionTestCase.php
@@ -2095,9 +2095,38 @@ abstract class FrameworkExtensionTestCase extends TestCase
         $this->assertEquals(new Reference('app.another_bus'), $container->getDefinition('mailer.mailer')->getArgument(1));
     }
 
-    public function testHttpClientMockResponseFactory()
+    public function testHttpClientMockResponseFactoryNoService()
     {
-        $container = $this->createContainerFromFile('http_client_mock_response_factory');
+        $container = $this->createContainerFromFile('http_client_scoped_mock_response_factory');
+
+        $this->assertFalse($container->hasDefinition('http_client.mock_client.notMocked'));
+
+        $definition = $container->getDefinition('http_client.mock_client.mocked');
+
+        $this->assertSame(MockHttpClient::class, $definition->getClass());
+        $this->assertCount(0, $definition->getArguments());
+    }
+
+    public function testHttpClientMockResponseFactoryWithService()
+    {
+        $container = $this->createContainerFromFile('http_client_scoped_mock_response_factory_service');
+        $this->assertFalse($container->hasDefinition('http_client.mock_client.notMocked'));
+
+        $definition = $container->getDefinition('http_client.mock_client.mocked');
+
+        $this->assertSame(MockHttpClient::class, $definition->getClass());
+        $this->assertCount(1, $definition->getArguments());
+
+        $argument = $definition->getArgument(0);
+
+        $this->assertInstanceOf(Reference::class, $argument);
+        $this->assertSame('http_client.transport', current($definition->getDecoratedService()));
+        $this->assertSame('my_response_factory', (string) $argument);
+    }
+
+    public function testHttpClientMockResponseFactoryService()
+    {
+        $container = $this->createContainerFromFile('http_client_mock_response_factory_service');
 
         $definition = $container->getDefinition('http_client.mock_client');
 
@@ -2109,6 +2138,16 @@ abstract class FrameworkExtensionTestCase extends TestCase
         $this->assertInstanceOf(Reference::class, $argument);
         $this->assertSame('http_client.transport', current($definition->getDecoratedService()));
         $this->assertSame('my_response_factory', (string) $argument);
+    }
+
+    public function testHttpClientUseMockResponseFactory()
+    {
+        $container = $this->createContainerFromFile('http_client_mock_response_factory');
+
+        $definition = $container->getDefinition('http_client.mock_client');
+
+        $this->assertSame(MockHttpClient::class, $definition->getClass());
+        $this->assertCount(0, $definition->getArguments());
     }
 
     public function testRegisterParameterCollectingBehaviorDescribingTags()