+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -51,6 +51,7 @@ CHANGELOG
 6.4
 ---
 
+ * Add support for setting mock_response_factory per scoped http client
  * Add `HttpClientAssertionsTrait`
  * Add `AbstractController::renderBlock()` and `renderBlockView()`
  * Add native return type to `Translator` and to `Application::reset()`